migrated utilities which are specific to sbml_dfs_core
